ENHANCEMENT: Refactored and moved functionality out of the Tracker() class to own sub-classes
